feat: move write blocks for client
